Fix recursive initialization of convars/concommands
'CSourceAppSystemGroup::PreInit' could be called multiple times. Added check for its actual stage before running any initialization code. This fixes the recursive init problem.

+++ r5dev/launcher/IApplication.cpp
@@ -38,11 +38,15 @@
 //-----------------------------------------------------------------------------
 bool CSourceAppSystemGroup::PreInit(CSourceAppSystemGroup* pSourceAppSystemGroup)
 {
-	ConVar::InitShipped();
-	ConVar::PurgeShipped();
-	ConCommand::StaticInit();
-	ConCommand::InitShipped();
-	ConCommand::PurgeShipped();
+	if (pSourceAppSystemGroup->GetCurrentStage() == CSourceAppSystemGroup::CREATION)
+	{
+		ConVar::InitShipped();
+		ConVar::PurgeShipped();
+		ConCommand::StaticInit();
+		ConCommand::InitShipped();
+		ConCommand::PurgeShipped();
+	}
+
 	return CSourceAppSystemGroup__PreInit(pSourceAppSystemGroup);
 }
